[client] Data instances uses enum DataType
