+++ python/rain/common/__init__.py
@@ -3,3 +3,4 @@ from .data_instance import DataInstance  # noqa
 from .errors import RainException, RainWarning  # noqa
 from .errors import SessionException, TaskException  # noqa
 from .ids import ID  # noqa
+from .datatype import DataType  # noqa
